requirements.txt
-i https://pypi.org/simple

asgiref==3.2.10; python_version >= '3.5'

cffi

cryptography==3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

django-filter==2.3.0

django-rest-knox==4.1.0

django==3.1.1

djangorestframework==3.11.1

pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pytz==2020.1

six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

